requirements.txt
fastapi
uvicorn
motor
pydantic
requests
python-dotenv
matplotlib
pytest
httpx
pytest-asyncio
